started matching functions in ftdonkey_1.s (#424)
* DOL Matching - finished Dk6.c, specialHi, and all of the functions in ft_donkey2.s

* started matching ftdonkey1 functions from the bottom, in reverse, these functions are separated from the rest of dk functions for some reason

* fixed a function

* created several dk files for functions from ftdonkey_1.s, working in reverse is a little odd but DOL is matching, this commit

* NONMatching, started next dk file, need to finish it for floats to match

* matched some tougher functions, DOL seems to be off only by floats

* finished matching all the functions in ftdonkey.s, needs cleanup, fix some fake stuff, and some of the files are probably ftwalkcommon.c files
