+++ include/functions.h
@@ -1221,7 +1221,7 @@ void ftAnim_SetAnimRate(struct _HSD_GObj*, f32);
 BOOL ftAnim_IsFramesRemaining(struct _HSD_GObj*);
 // ? func_8006F368(?);
 // ? func_8006F3DC(?);
-// ? func_8006F484(?);
+f32 func_8006F484(void*); 
 // ? func_8006F4C8(?);
 // ? func_8006F628(?);
 // ? func_8006F7C8(?);
@@ -1832,7 +1832,7 @@ void func_8008D930();
 // ? func_8008DA4C(?);
 // ? func_8008DB10(?);
 // ? func_8008DC0C(?);
-// ? func_8008DCE0(?);
+void func_8008DCE0(HSD_GObj*, s32, f32);
 // ? func_8008E498(?);
 // ? func_8008E5A4(?);
 void func_8008E908(struct _HSD_GObj*, f32);
@@ -2417,7 +2417,7 @@ void func_800C9034();
 // ? func_800C9528(?);
 // ? func_800C97A8(?);
 // ? func_800C97DC(?);
-// ? func_800C9840(?);
+void func_800C9840(HSD_GObj*, s32, s32, f32, f32, f32);
 // ? func_800C98AC(?);
 // ? func_800C9924(?);
 // ? func_800C9C2C(?);
@@ -2441,7 +2441,7 @@ s32 func_800CAE80();
 // ? func_800CAED0(?);
 // ? func_800CAF78(?);
 // ? func_800CB024(?);
-// ? func_800CB110(?);
+void func_800CB110(HSD_GObj*, s32, f32);
 // ? func_800CB250(?);
 // ? func_800CB334(?);
 // ? func_800CB438(?);
@@ -2557,7 +2557,7 @@ void func_800D3158(struct _HSD_GObj* fighterObj);
 void func_800D4F24(struct _HSD_GObj*, s32);
 // ? func_800D4FF4(?);
 // ? func_800D5600(?);
-// ? func_800D5AEC(?);
+void func_800D5AEC(HSD_GObj*, s32, s32, s32, Fighter*, f32, f32);
 // ? func_800D5BF8(?);
 void func_800D5CB0(struct _HSD_GObj*, s32, f32);
 // ? func_800D5D3C(?);
@@ -2709,27 +2709,27 @@ void func_800DF0D0();
 // ? func_800DF8B0(?);
 // ? func_800DF8E8(?);
 // ? func_800DF910(?);
-// ? func_800DF938(?);
-// ? func_800DF980(?);
-// ? func_800DFA70(?);
-// ? func_800DFA98(?);
-// ? func_800DFAE4(?);
-// ? func_800DFBF8(?);
-// ? func_800DFC70(?);
-// ? func_800DFCA4(?);
-// ? func_800DFDDC(?);
-// ? func_800DFEC8(?);
-// ? func_800E0060(?);
-// ? func_800E0134(?);
-// ? func_800E017C(?);
-// ? func_800E0294(?);
-// ? func_800E0350(?);
-// ? func_800E0378(?);
-// ? func_800E03C0(?);
-// ? func_800E04A4(?);
-// ? func_800E06D8(?);
-// ? func_800E0750(?);
-// ? func_800E07B0(?);
+// ? ftDonkey_800DF938(?);
+// ? ftDonkey_800DF980(?);
+// ? ftDonkey_800DFA70(?);
+// ? ftDonkey_800DFA98(?);
+// ? ftDonkey_800DFAE4(?);
+// ? ftDonkey_GetWalkType_800DFBF8(?);
+// ? ftDonkey_800DFC70(?);
+// ? ftDonkey_800DFCA4(?);
+// ? ftDonkey_800DFDDC(?);
+// ? ftDonkey_800DFEC8(?);
+// ? ftDonkey_800E0060(?);
+// ? ftDonkey_800E0134(?);
+// ? ftDonkey_800E017C(?);
+// ? ftDonkey_800E0294(?);
+// ? ftDonkey_800E0350(?);
+// ? ftDonkey_800E0378(?);
+// ? ftDonkey_800E03C0(?);
+// ? ftDonkey_800E04A4(?);
+// ? ftDonkey_800E06D8(?);
+// ? ftDonkey_800E0750(?);
+// ? ftDonkey_800E07B0(?);
 // ? ftMario_OnLoadForDrMario(?);
 // ? ftMario_LoadSpecialAttrs(?);
 // ? func_800E0D1C(?);
@@ -3003,7 +3003,7 @@ void func_800F5C34();
 // ? ftDonkey_8010E464(?);
 // ? ftDonkey_8010E4EC(?);
 // ? ftDonkey_DestroyAllEffectsPlus(?);
-// ? func_80110074(?);
+// ? ftDonkey_DestroyAllEffects_SpecialHi(?);
 // ? func_80110490(?);
 // ? func_80110610(?);
 // ? func_80110788(?);
